requirements.txt
discord.py
aiohttp
colorama
beautifulsoup4